refactor: Move selector handling to own file
